requirements.txt
streamlit
Pillow
torch>=1.7.0,<1.11  
torchvision  
ultralytics>=6.0  